Include est. expended in Obligated Funds graph
To conform to new designs, update the bar graph in the obligated funds
section of the reporting page to include the idea of "estimated"
expended funds.

In addition to markup and CSS changes, this PR also added an
"estimated_expended_funds" property to some of the dictionaries retured
by mock CSP reporting functions.