WIP - refactor reporting data
This branch is a start at refactoring how demo data for reports gets
ingested and processed by ATAT

All of the reporting data was moved to a JSON file. The concept of
what JEDI CLIN a particular environment drew money from was added to the
data.

This also involved starting to refactor some of the reporting class
methods, which is where this commit left off.

I'm not sure storing all of the fixture data in one file is the right
approach -- because the method to get JEDI seems like it will be tedious
to implement, it might be better to make a few fixture data files that
represent what mock API responses would look like -- one for app / env
spending, one for JEDI CLIN spending, etc. We would still coordinate the
values between the fixure files to ensure that they "add up" on the
reporting page.

Finally, I'm not sure how we'd incorporate "amount unspent" for expired
task orders, unless that data was added to the seed_sample script.